+++ scripts/BoonsAndBurdens/backgrounds_player_merged/infected.lua
@@ -2,7 +2,6 @@
 ---@diagnostic disable: assign-type-mismatch
 ---@diagnostic disable: undefined-field
 local I = require("openmw.interfaces")
--- local self = require("openmw.self")
 local core = require("openmw.core")
 
 local vampireWindow = require("scripts.BoonsAndBurdens.ui.infected")